refactor: kafka consumer를 infra layer에서 application layer로 이동
