+++ src/main/java/kr/hhplus/be/server/application/dataplatform/messaging/DataPlatformKafkaConsumer.java
@@ -1,11 +1,10 @@
-package kr.hhplus.be.server.infrastructure.messaging.dataplatform;
+package kr.hhplus.be.server.application.dataplatform.messaging;
 
+import kr.hhplus.be.server.domain.dataplatform.service.DataPlatformSendService;
 import kr.hhplus.be.server.domain.payment.event.PaymentCompletedEvent;
-import kr.hhplus.be.server.infrastructure.client.dataplatform.DataPlatformClient;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.kafka.annotation.KafkaListener;
-import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
 
 import static kr.hhplus.be.server.infrastructure.messaging.config.KafkaConstance.DATA_PLATFORM_CONSUMER_GROUP;
@@ -15,16 +14,16 @@ import static kr.hhplus.be.server.infrastructure.messaging.config.KafkaConstance
 @Component
 @RequiredArgsConstructor
 public class DataPlatformKafkaConsumer {
-    private final DataPlatformClient dataPlatformClient;
+    private final DataPlatformSendService dataPlatformSendService;
 
     @KafkaListener(
         topics = PAYMENT_COMPLETED_TOPIC,
         groupId = DATA_PLATFORM_CONSUMER_GROUP
     )
-    @Async
     public void consume(PaymentCompletedEvent event) {
+        // event -> DTO 변환 (생략)
         log.info("[DATA-PLATFORM-SERVICE][DataPlatformKafkaConsumer] 결제 데이터 전송: 결제 ID - {}", event.paymentId());
-        dataPlatformClient.sendData();
+        dataPlatformSendService.sendReservationPaymentResult();
         log.info("[DATA-PLATFORM-SERVICE][DataPlatformKafkaConsumer] 결제 데이터 전송 완료");
     }
 }
