+++ src/main/java/kr/hhplus/be/server/domain/dataplatform/client/DataPlatformSender.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.domain.dataplatform.client;
+
+public interface DataPlatformSender {
+    Boolean sendData();
+}
