+++ src/main/resources/application.yml
@@ -68,9 +68,3 @@ springdoc:
 
 #---
 #spring.config.activate.on-profile: local, test
-#
-#spring:
-#  datasource:
-#    url: jdbc:mysql://localhost:3306/hhplus?characterEncoding=UTF-8&serverTimezone=UTC
-#    username: application
-#    password: application
